# sera-back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sera API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/projects/{projectId}/videos": {
            "get": {
                "tags": [
                    "Video Review"
                ],
                "summary": "Get all video reviews of a project",
                "operationId": "fcd490f02968989ee213153dc79ebc27",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "Id of the project",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "version",
                        "in": "query",
                        "description": "Version of the video review",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Video reviews found"
                    },
                    "404": {
                        "description": "No video reviews found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Video Review"
                ],
                "summary": "Create a video review",
                "operationId": "32aeabc020ba14054c0ab65b08fe400d",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "Id of the project",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Video review object that needs to be added to the project",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "provider": {
                                        "type": "string",
                                        "example": "html5"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "video"
                                    },
                                    "resolution": {
                                        "type": "string",
                                        "example": "1080"
                                    },
                                    "url": {
                                        "type": "string",
                                        "example": "/test.mkv"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "test"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "test"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Video review created"
                    },
                    "404": {
                        "description": "Project not found"
                    }
                }
            }
        },
        "/api/videos/{version}": {
            "delete": {
                "tags": [
                    "Video Review"
                ],
                "summary": "Delete a video review",
                "operationId": "621be293d9a3b1e349ab265495f4e2d3",
                "parameters": [
                    {
                        "name": "version",
                        "in": "path",
                        "description": "Version of the video review",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Video review deleted"
                    },
                    "404": {
                        "description": "Video review not found"
                    }
                }
            }
        },
        "/api/projects/{projectId}/videos/{version}": {
            "post": {
                "tags": [
                    "Video Review"
                ],
                "summary": "Add a comment to a video review",
                "operationId": "749dc5fbfb87eaa226b835bfc89c2f5a",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "Id of the project",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "version",
                        "in": "path",
                        "description": "Version of the video review",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Comment object that needs to be added to the video review",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "message": {
                                        "type": "string",
                                        "example": "test"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Comment created"
                    },
                    "404": {
                        "description": "Video review not found"
                    }
                }
            }
        },
        "/api/projects/{projectId}/videos/getuploadurl": {
            "get": {
                "tags": [
                    "Video Review"
                ],
                "summary": "Get a temporary upload url for a video review",
                "operationId": "3220387cb8d441455c9217516b75140a",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "Id of the project",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Temporary upload url found"
                    },
                    "401": {
                        "description": "User not connected"
                    },
                    "404": {
                        "description": "Project not found"
                    }
                }
            }
        },
        "/api/projects/{projectId}/videos/validated": {
            "get": {
                "tags": [
                    "Video Review"
                ],
                "summary": "Get the validated video review of a project",
                "operationId": "154d11a501fe0ba16c38cf4d1bff08d3",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "Id of the project",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Video review found"
                    },
                    "404": {
                        "description": "Video review not found"
                    }
                }
            }
        }
    }
}